move apply to it's own package
